Fix enzyme testing (#1)
* Patch Enzyme testing issue with missing SVG methods

This is a patch for this issue:

https://github.com/moarwick/react-mt-svg-lines/issues/11

JSDom has very little support for SVG, and so calls to many SVG element methods will fail during Enzyme testing with `mount`.

This fork patches the issue by testing for the existence of getTotalLength, and simlply returning 0 if `undefined`. Not ideal, but allows components utilising this lib to be tested.

* Produce new dist

* Move react-hot-loader to dependencey list

The seems to require this during testing. No idea why